Weaken history invariant to account for resets
When a history is reset, we only adjust offsets into a ring buffer,
leaving the old values in the ring buffer intact. This causes the
history invariant to be temporarily invalided. This is okay since
the duration is set to zero, and so the rank is zero, too. As soon
as a history is modified (start/stop), the invariant is restored.